+++ src/Stormpath.Owin.Middleware/ViewModelBuilder/ForgotPasswordFormViewModelBuilder.cs
@@ -1,7 +1,8 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using Stormpath.Owin.Abstractions.Configuration;
-namespace Stormpath.Owin.Abstractions.ViewModel
+using Stormpath.Owin.Abstractions.ViewModel;
+
+namespace Stormpath.Owin.Middleware.ViewModelBuilder
 {
     public sealed class ForgotPasswordFormViewModelBuilder
     {
@@ -22,12 +23,9 @@ namespace Stormpath.Owin.Abstractions.ViewModel
             var result = new ForgotPasswordFormViewModel(baseViewModelBuilder.Build());
 
             // Add a state (CSRF) token
-            // Add a state (CSRF) token
-            throw new NotImplementedException("TODO");
+            result.StateToken = new StateTokenBuilder(_configuration.OktaEnvironment.ClientSecret).ToString();
 
-            //result.StateToken = new StateTokenBuilder(_configuration.Client.ApiKey).ToString();
-
-            //return result;
+            return result;
         }
     }
 }