+++ changelog.md
@@ -5,4 +5,4 @@ Migration information TODO.
 ## Breaking changes
 
 * The `StateTokenBusilder` and `StateTokenParser` classes were moved from `Stormpath.Owin.Abstractions` to `Stormpath.Owin.Middleware`.
-* The `LoginFormViewModelBuilder` class was moved from `Stormpath.Owin.Abstractions` to `Stormpath.Owin.Middleware`.
+* The `*FormViewModelBuilder` classes was moved from `Stormpath.Owin.Abstractions` to `Stormpath.Owin.Middleware`.